Delete socket file when server isn't running (#10)
Sometimes the server dies without removing the socket file. When this
happens the server will fail to start even though there is no running
server.

Instead of failing immediately, the server command now attempts to talk
to the server running on the socket. If it succeeds, it exits since a
server is already running. If it fails, it removes the socket file and
starts the server.