Add support for linux clipboard using xclip command
Extracted most of macosClipboard into commandClipboard, which as a
platform-agnostic class that can handle any copy/paste command.

Added build-tagged files with build constraints for darwin and linux
respectively to specify pbcopy/pbpaste and xclip respectively.

Linux has multiple clipboards but I've chosen to use the one mapped to
Ctrl-C / Ctrl-V. Possibly in the future this could be configurable.