Support linux (#17)
This adds support for running rdm server on linux machines by replacing macos-specific clipboard and URL-opening commands with an abstraction and using build constraints to select the correct implementation. Specifically pbcopy/pbpaste becomes xclip and open becomes xdg-open. One advantage of this approach is that the binary now only builds on platforms where it is capable of running.