Update to universal logger.
This consolidates everything under a single logger which is unprefixed
and delivered to stderr when being used for human-readable messages (ie
CLI context) and timestamp-prefixed and multiplexed to both the log file and
stdout when used in server context. We manage the transition from CLI
to long-lived process by reconfiguring the existing logger built in
main.go. One advantage of this is that things that need to log inside
main.go but the server log should know about (like signal receipt) get
sent to the right place instead of a now-likely-disconnected stderr.